# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aldeias Games API",
    "description": "API para plataforma SaaS multi-tenant de angariacao de fundos em comunidades locais portuguesas",
    "version": "3.12.0",
    "contact": {
      "name": "Sandro Pereira",
      "url": "https://github.com/smpsandro1239/Aldeias_Games"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desenvolvimento"
    },
    {
      "url": "https://aldeias-games.vercel.app",
      "description": "Producao"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_token"
      }
    },
    "schemas": {
...
      "Jogo": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "nome": { "type": "string" },
          "tipo": { "type": "string", "enum": ["rifa", "poio_da_vaca", "euromilhoes", "raspadinha"] },
          "preco": { "type": "number" },
          "stockInicial": { "type": "integer" },
          "stockAtual": { "type": "integer" },
          "isSorteado": { "type": "boolean" },
          "isFinalizado": { "type": "boolean" },
          "aldeiaId": { "type": "string" }
        }
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] },
    { "cookieAuth": [] }
  ],
  "paths": {
...
    "/api/auth/2fa": {
      "get": {
        "tags": ["2FA"],
        "summary": "Verificar estado 2FA",
        "description": "Retorna se 2FA esta ativo para o utilizador atual.",
        "responses": {
          "200": {
            "content": { "application/json": { "schema": { "type": "object", "properties": { "enabled": { "type": "boolean" } } } } }
          },
          "401": { "description": "Nao autorizado" }
        }
      },
      "post": {
        "tags": ["2FA"],
        "summary": "Configurar/verificar/desativar 2FA",
        "description": "Acoes: setup (gerar QR), verify (ativar), disable (desativar)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "action": { "type": "string", "enum": ["setup"] }
                    }
                  },
                  {
                    "type": "object",
                    "required": ["code"],
                    "properties": {
                      "action": { "type": "string", "enum": ["verify"] },
                      "code": { "type": "string", "description": "Codigo TOTP de 6 digitos" }
                    }
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": { "type": "string", "enum": ["disable"] }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Operacao bem-sucedida" },
          "400": { "description": "Codigo invalido ou 2FA nao configurado" },
          "429": { "description": "Rate limit excedido (5 tentativas/5min)" }
        }
      }
    },
...
    "/api/jogos": {
      "get": {
        "tags": ["Jogos"],
        "summary": "Listar jogos",
        "security": [],
        "responses": {
          "200": {
            "content": { "application/json": { "schema": { "type": "object", "properties": { "data": { "type": "array", "items": { "$ref": "#/components/schemas/Jogo" } } } } } }
          }
        }
      },
      "post": {
        "tags": ["Jogos"],
        "summary": "Criar jogo (admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["nome", "tipo", "preco", "stockInicial"],
                "properties": {
                  "nome": { "type": "string" },
                  "tipo": { "type": "string", "enum": ["rifa", "poio_da_vaca", "euromilhoes", "raspadinha"] },
                  "preco": { "type": "number" },
                  "stockInicial": { "type": "integer" },
                  "descricao": { "type": "string" },
                  "dataInicio": { "type": "string", "format": "date-time" },
                  "dataFim": { "type": "string", "format": "date-time" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Jogo criado" },
          "400": { "description": "Dados invalidos" }
        }
      }
    },
    "/api/jogos/{id}": {
      "get": {
        "tags": ["Jogos"],
        "summary": "Obter jogo por ID",
        "security": [],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Jogo" } } } },
          "404": { "description": "Jogo nao encontrado" }
        }
      }
    },
...
  }
}